Actualiza

+++ test/dummy/db/structure.sql
@@ -1133,8 +1133,8 @@ CREATE TABLE public.sipd_dominio_operaen_pais (
 --
 
 CREATE TABLE public.sipd_dominio_usuario (
-    dominio_id integer NOT NULL,
-    usuario_id integer NOT NULL
+    dominio_id bigint NOT NULL,
+    usuario_id bigint NOT NULL
 );
